+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/UserStorage.java
@@ -7,9 +7,14 @@ import java.util.Map;
 
 public interface UserStorage {
     Collection<User> getAllUser();
+
     User createUser(User user);
+
     User updateUser(User user);
-    User findUserById(Integer id);
-    Map<Integer, User> getUsers();
+
+    User findUserById(Long id);
+
+    Map<Long, User> getUsers();
+
     long getGeneratorId();
 }
